+++ src/components/LoanFinding/QuickFiltersSection.vue
@@ -14,27 +14,54 @@
 			@reset-filters="resetFilters"
 			@handle-overlay="handleQuickFiltersOverlay"
 		/>
+		<div
+			class="tw-grid tw-grid-cols-1 md:tw-grid-cols-2 lg:tw-grid-cols-3 tw-gap-4 tw-mt-2"
+		>
+			<kiva-classic-basic-loan-card
+				v-for="(loan, index) in loans"
+				:key="`loan-${loan.id}`"
+				:item-index="index"
+				:loan-id="loan.id"
+				:show-action-button="true"
+			/>
+		</div>
+		<div class="tw-w-full tw-my-4">
+			<kv-pagination
+				:total="totalCount"
+				:limit="loanSearchState.pageLimit"
+				:offset="loanSearchState.pageOffset"
+				@page-changed="pageChange"
+			/>
+		</div>
 	</div>
 </template>
 
 <script>
 import QuickFilters from '@/components/LoansByCategory/QuickFilters/QuickFilters';
-import { runFacetsQueries, fetchLoanFacets } from '@/util/loanSearch/dataUtils';
+import { runFacetsQueries, fetchLoanFacets, runLoansQuery } from '@/util/loanSearch/dataUtils';
 import { fetchCategories, FLSS_ORIGIN_CATEGORY } from '@/util/flssUtils';
 import { transformIsoCodes } from '@/util/loanSearch/filters/regions';
+import KivaClassicBasicLoanCard from '@/components/LoanCards/KivaClassicBasicLoanCard';
+import KvPagination from '@/components/Kv/KvPagination';
 
 export default {
 	name: 'QuickFiltersSection',
 	components: {
-		QuickFilters
+		QuickFilters,
+		KivaClassicBasicLoanCard,
+		KvPagination
 	},
 	inject: ['apollo'],
 	data() {
 		return {
-			filtersLoaded: false,
+			totalCount: 0,
 			targetedLoanChannelURL: '',
+			filtersLoaded: false,
 			flssLoanSearch: {},
-			loanSearchState: {},
+			loanSearchState: {
+				pageOffset: 0,
+				pageLimit: 6
+			},
 			loans: [],
 			quickFiltersOptions: {
 				categories: [{
@@ -56,14 +83,40 @@ export default {
 	async mounted() {
 		this.allFacets = await fetchLoanFacets(this.apollo);
 		await this.fetchFilterData(this.flssLoanSearch);
-	},
-	computed: {
-		totalCount() {
-			return 0;
-		},
+		const { loans, totalCount } = await runLoansQuery(
+			this.apollo,
+			{ ...this.flssLoanSearch, ...this.loanSearchState },
+			FLSS_ORIGIN_CATEGORY
+		);
+		this.loans = loans;
+		this.totalCount = totalCount;
 	},
 	methods: {
-		updateQuickFilters() {
+		async updateQuickFilters(filter) {
+			if (filter.gender !== undefined) {
+				this.flssLoanSearch.gender = filter.gender;
+			} else if (filter.sortBy) {
+				this.flssLoanSearch.sortBy = filter.sortBy;
+			} else if (filter.country) {
+				this.flssLoanSearch.countryIsoCode = filter.country;
+			} else {
+				// We want to reset the flss paramaters for categories
+				delete this.flssLoanSearch.sectorId;
+				delete this.flssLoanSearch.tagId;
+				delete this.flssLoanSearch.activityId;
+				delete this.flssLoanSearch.themeId;
+				this.flssLoanSearch = {
+					...this.flssLoanSearch,
+					...filter
+				};
+			}
+			const { loans, totalCount } = await runLoansQuery(
+				this.apollo,
+				{ ...this.flssLoanSearch, ...this.loanSearchState },
+				FLSS_ORIGIN_CATEGORY
+			);
+			this.loans = loans;
+			this.totalCount = totalCount;
 		},
 		resetFilters() {
 		},
@@ -138,6 +191,9 @@ export default {
 			];
 
 			this.filtersLoaded = true;
+		},
+		pageChange({ pageOffset }) {
+			this.loanSearchState.pageOffset = pageOffset;
 		},
 	},
 };